# Fixed Date time stamps so all are in the correct format and now read as PosixCT.
Identified that there were 3 PIT tag series and was able to merge the PIT tag data from the livewell scanner to the JSATS database using the last 5 digits of the PIT tag number from the livewell scanner.

Corrected release location names to their appropriate spelling

Identified one fish that was scanned at the livewells but did not have an associated acoustic tag.

Wrote csv file for my own records and a txt file for ODFW